Add files via upload
I created this script after much trouble finding exactly the right instructions to install Kiwix-Serve on a current version of Raspberry Pi OS. After much trial and error, I was able to get everything installed and working, by using several handsfull of tutorials on the subject. I created this script in the hopes that some future person may be saved the troubles I faced. I tried to encompass every feature one could want in a manager for the Kiwix-Serve application, but if I missed anything, or if you'd like to see anything changed or added, please let me know.